Deprecate `hmn` and `lgc` modules from the `node_classification` package (#5166)
* Reorganize tests.

Condense all node-classification tests into single test module
one level up.

* Add tests for deprecation warnings.

* Prevent test crosstalk.

* Refactor pkg with __getattr__ and __dir__.

Add a deprecation warning whenever the hmn or lgc attributes
are accessed.

* Add deprecation reminder to developer docs.

* Add release note.
