+++ networkx/algorithms/node_classification/__init__.py
@@ -18,5 +18,35 @@ For example:
 
 """
 
-from networkx.algorithms.node_classification.hmn import *
-from networkx.algorithms.node_classification.lgc import *
+
+def __getattr__(name):
+    if name in ("hmn", "lgc"):
+        import warnings
+        import importlib
+
+        fn_name = (
+            "harmonic_function" if name == "hmn" else "local_and_global_consistency"
+        )
+        msg = (
+            f"The {name}  module is deprecated and will be removed in version 3.0.\n"
+            f"Access `{fn_name}` directly from `node_classification`:\n\n"
+            "    from networkx.algorithms import node_classification\n"
+            f"    node_classification.{fn_name}\n"
+        )
+        warnings.warn(msg, category=DeprecationWarning, stacklevel=2)
+        return importlib.import_module(
+            f".{name}", "networkx.algorithms.node_classification"
+        )
+    if name == "harmonic_function":
+        from .hmn import harmonic_function
+
+        return harmonic_function
+    if name == "local_and_global_consistency":
+        from .lgc import local_and_global_consistency
+
+        return local_and_global_consistency
+    raise AttributeError(f"module {__name__} has no attribute {name}")
+
+
+def __dir__():
+    return ["harmonic_function", "local_and_global_consistency"]
